Remove prebuilt com.android.art APEXes for the time being.
We are not currently targeting building platform with ART prebuilts in
AOSP.

This avoids ambiguity when deciding which prebuilt APEX to use for
deapexing in platform builds where there are other prebuilts than
these. A proper solution for that requires introducing a product
variable to specify which APEXes to use, which requires more
engineering effort (b/192006406).

Test: m droid SOONG_CONFIG_art_module_source_build=false
  on internal master
Test: art/build/build-art-module.sh
  on master-art
Test: art/tools/buildbot-build.sh --target
  on master-art
Bug: 192006406
Bug: 202192894
Change-Id: I342b9f400e9778675e21103428d65feb5d4640f3